float

+++ app.py
@@ -20,8 +20,8 @@ def index():
 def predict():
     # DATE,TMAX,TMIN,RAIN for seattle
     date = request.args.get("att0", "")
-    t_Max = request.args.get("att1", type=int)
-    t_Min = request.args.get("att2", type=int)
+    t_Max = request.args.get("att1", type=float)
+    t_Min = request.args.get("att2", type=float)
     # rain = request.args.get("att3", "")
 
     prediction = predict_rain_well([date, t_Max, t_Min])
